+++ src/Sharpbrake.Client/Properties/AssemblyInfo.cs
@@ -11,9 +11,9 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCompany("Airbrake Technologies, Inc.")]
 [assembly: AssemblyCopyright("Copyright (c) 2017 Airbrake Technologies, Inc.")]
 
-[assembly: AssemblyVersion("3.0.3.0")]
-[assembly: AssemblyFileVersion("3.0.3.0")]
-[assembly: AssemblyInformationalVersion("3.0.3.0")]
+[assembly: AssemblyVersion("3.1.0.0")]
+[assembly: AssemblyFileVersion("3.1.0.0")]
+[assembly: AssemblyInformationalVersion("3.1.0.0")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
 // to COM components.  If you need to access a type in this assembly from