+++ CHANGELOG.md
@@ -3,8 +3,12 @@ Sharpbrake Changelog
 
 ### master
 
+### [v3.1.0][v3.1.0] (May 20, 2017)
+
 * Started sending error severity (defaults to `error`)
   ([#55](https://github.com/airbrake/sharpbrake/pull/55))
+* Added support for notice truncation
+  ([#54](https://github.com/airbrake/sharpbrake/pull/54))
 
 ### [v3.0.3][v3.0.3] (April 21, 2017)
 
@@ -18,14 +22,15 @@ Sharpbrake Changelog
 
 ### [v3.0.1][v3.0.1] (March 15, 2017)
 
-* Maintenance updates.
+* Maintenance updates
 
 ### [v3.0.0][v3.0.0] (March 14, 2017)
 
 * Version 3 is written from scratch. See [here](https://github.com/airbrake/sharpbrake#key-features)
-  for new features that have been introduced.
+  for new features that have been introduced
 
 [v3.0.0]: https://github.com/airbrake/sharpbrake/releases/tag/v3.0.0
 [v3.0.1]: https://github.com/airbrake/sharpbrake/releases/tag/v3.0.1
 [v3.0.2]: https://github.com/airbrake/sharpbrake/releases/tag/v3.0.2
 [v3.0.3]: https://github.com/airbrake/sharpbrake/releases/tag/v3.0.3
+[v3.1.0]: https://github.com/airbrake/sharpbrake/releases/tag/v3.1.0